+++ CHANGELOG.md
@@ -67,7 +67,7 @@
 ### Added
 - Toolbar: new preset 'Statust bar' available for empty toolbars. This preset creates a statust bar similar to the one found at DUI/CUI at bottom, but it may be tweaked with any TF (even for the volume or selection size fields), like the 'display_tf' button (see below).
 - Toolbar: new entry on toolbar menu to share current UI settings across all toolbar panels on foobar2000. See [here](https://hydrogenaud.io/index.php/topic,120978.msg1061005.html#msg1061005).
-- Other tools\Group Tagger: new tool to calculate average, sum or count of tag values for selected tracks and -optionally- save the result to another tag. It may be used to calculate album ratings, weights by duration, etc. See [here](https://hydrogenaudio.org/index.php/topic,128376.new.html).
+- Other tools\Group Tagger: new tool to calculate average, sum, count or mode of tag values for selected tracks and -optionally- save the result to another tag. It may be used to calculate album ratings, weights by duration, etc. See [here](https://hydrogenaudio.org/index.php/topic,128376.new.html).
 - Pools: added 'Radio (Alistral)' pools presets, based on [Alistral](https://github.com/RustyNova016/Alistral?tab=readme-ov-file#radio) radio algorithms (see link for more info). Restore defaults to load the new presets.
 - Pools: expanded 'Genre mixes' presets. Restore defaults at pools submenu to show them.
 - Pools: expanded 'Top tracks (...)' presets. Restore defaults at pools submenu to show them.